% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{boot.pair}
\alias{boot.pair}
\title{Compare dependent pairwise kappas (bootstrap method)}
\usage{
boot.pair(cluster_id, data, weight = "equal", a.level = 0.05, ITN = 1000,
  summary_k = T)
}
\arguments{
\item{cluster_id}{a vector of lenght N with the identification number of the clusters}

\item{data}{a N x R matrix representing the classification of the N items by the R observers. The kappa coefficients are computed between column (1,2), (3,4), etc....}

\item{weight}{the weighting scheme to be used for kappa coefficients. 'unweighted' for Cohen's kappa, 'equal' for linear weights and 'squared' for quadratic weights}

\item{a.level}{significance level}

\item{ITN}{the number of bootstrap iterations}

\item{summary_k, }{if true, Hotteling's T square test is performed, if false, only the bootstraped kappa coefficients are returned}
}
\value{
$kappa a G x 2 matrix with the G kappa coefficients to be compared in the first column and their corresponding standard error in the second column

$T_test a vector of length 2 with the value of Hotelling's T square test as first element and the corresponding p-value as second element

$confidence confidence intervals for the pairwise comparisons of the kappa coefficients

$cor the G x G correlation matrix of the kappa coefficients

$K when summary_k is false, the ITN x G matrix with the bootstrapped kappa coefficients
}
\description{
This function performs Hotelling's T square test using a variance-covariance matrix based on the bootstrap method to compare dependent pairwise kappa coefficients
}
\details{
This function compares several dependent pairwise kappa coefficients using Hotelling's T square with the variance-covariance matrix obtained by the bootstrap method. If only one kappa is computed, it returns the estimate and confidence interval.
}
\examples{
 
#dataset (not multilevel) (Vanbelle and Albert, 2008)

set.seed(103) #to get the same results as in the paper
data(depression)
attach(depression)
a<-boot.pair(data=cbind(diag,BDI,diag,GHQ),cluster_id=ID,weight='unweighted')


#dataset (multilevel) (Vanbelle, xxx)

data(FEES)
attach(FEES)
dat<-cbind(val_CO,val_COR,val_MH,val_MHR,val_TB,val_TBR) #formating the data matrix
boot.pair(data=dat,cluster_id=subject,weight='equal',summary_k=FALSE)

}
\references{
Vanbelle S. and Albert A. (2008). A bootstrap method for comparing correlated kappa coefficients.  Journal of Statistical Computation and  Simulation, 1009-1015

Vanbelle S. Comparing dependent agreement coefficients obtained on multilevel data. submitted

Vanbelle S. (2014) A New Interpretation of the Weighted Kappa Coefficients. Psychometrika. Advance online publication.  doi: 10.1007/s11336-014-9439-4
}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
