% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{cov2cor}
\alias{cov2cor}
\title{convert a variance-covariance matrix in a correlation matrix}
\usage{
cov2cor(V)
}
\arguments{
\item{V}{a G x G matrix}
}
\value{
a G x G correlation matrix
}
\description{
convert a variance-covariance matrix in a correlation matrix
}
\details{
This function covert a variance-covariance matrix in a correlation matrix
}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
