% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\docType{data}
\name{CRACKLES}
\alias{CRACKLES}
\title{CRACKLES (Tromso study)}
\format{A data frame with 120 rows and 31 variables}
\description{
A dataset containing the assessment of the presence of crackles in 20 subjects by 28 observers. The variables are as follows:
\itemize{
  \item EXP1...4. Classification made by experts 1...4
  \item NOR1...4. Classification made by general practitioners 1...4 from Norway 
  \item RUS1...4. Classification made by general practitioners 1...4 from Russia 
  \item WAL1...4. Classification made by general practitioners 1...4 from Wales 
  \item NLD1...4. Classification made by general practitioners 1...4 from The Netherlands 
  \item PUL1...4. Classification made by pulmonologists 1...4
  \item STU1...4. Classification made by students 1...4
  \item patient. subject ID (1--20)
  \item UP. indicator variable. 1: examination of the upper posterior thorax; 0: otherwise
  \item LO. indicator variable. 1: examination of the lower posterior thorax; 0: otherwise  
}
}
