% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{delta.pair}
\alias{delta.pair}
\title{Compare dependent pairwise kappas (delta method)}
\usage{
delta.pair(data, cluster_id, weight, multilevel = T, a.level = 0.05)
}
\arguments{
\item{data}{a N x R matrix representing the classification of the N items by the R observers. The kappa coefficients are computed pairwise between column (1,2), (3,4), etc....}

\item{cluster_id}{a vector of lenght N with the identification number of the K clusters}

\item{weight}{the weighting scheme to be used in the computation of the kappa coefficients: 'unweighted' for Cohen's kappa, 'equal' for linear weights and 'squared' for quadratic weights}

\item{multilevel}{a binary indicator equal to TRUE in the presence of multilevel data and FALSE otherwise}

\item{a.level}{the significance level}
}
\value{
$kappa a G x 2 matrix with the G kappa coefficients to be compared in the first column and their corresponding standard error in the second column

$T_test a vector of length 2 with the value of Hotelling's T square test as first element and the p-value as second element

$confidence confidence intervals for the pairwise comparisons of kappa coefficients

$var the G x G correlation matrix of the kappa coefficients
}
\description{
This function performs Hotelling's T square test using a variance-covariance matrix based on the delta method to compare dependent pairwise kappa coefficients
}
\details{
This function compare several dependent kappa coefficients obtained between pairs of observers. It uses Hotelling's T square test with the variance-covariance matrix obtained by the delta method. If only a single kappa coefficient is computed, the kappa coefficient and its standard error are returned.
}
\examples{
 
#dataset (not multilevel) (Vanbelle and Albert, 2008)

data(depression)
attach(depression)
delta.pair(data=cbind(diag,BDI,diag,GHQ),cluster_id=ID,weight='unweighted',multilevel=FALSE)


#dataset (multilevel) (Vanbelle, xxx)

data(FEES)
attach(FEES)
dat<-cbind(val_CO,val_COR,val_MH,val_MHR,val_TB,val_TBR) #formating the data matrix
delta.pair(data=dat,cluster_id=subject,weight='equal')
}
\references{
Vanbelle S. and Albert A. (2008). A bootstrap method for comparing correlated kappa coefficients.  Journal of Statistical Computation and  Simulation, 1009-1015

Vanbelle S. (in press). Comparing dependent agreement coefficients obtained on multilevel data. Biometrical journal. doi: 10.1002/bimj.201600093

Vanbelle S. (2014). A New Interpretation of the Weighted Kappa Coefficients. Psychometrika. Advance online publication.  doi: 10.1007/s11336-014-9439-4
}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
