% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\docType{data}
\name{depression}
\alias{depression}
\title{Diagnosis of depression}
\format{A data frame with 50 rows and 4 variables}
\description{
A dataset containing the diagnosis of depression of 50 subjects on three binary scales: diagnosis, BDI and GHQ.  The variables are as follows:
\itemize{
  \item ID. subject ID (1--50)
  \item diag. depression diagnosis (0: no, 1: yes)
  \item BDI. depression diagnosis following BDI scale (0: no, 1: yes)
  \item GHQ. depression diagnosis following PHQ scale (0: no, 1: yes)
}
}
