% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\docType{data}
\name{FEES}
\alias{FEES}
\title{FEES}
\format{A data frame with 40 rows and 8 variables}
\description{
A dataset containing the valleculae score of 20 subjects obtained with thin and thick liquid.  The variables are as follows:
\itemize{
  \item subject. subject ID (1--37)
  \item swallow. type of swallowing (1=thin liquid, 2=thick liquid)
  \item val_TB. valleculae score given by rater TB on occasion  1 (0: no, 1: <50\%, 2: >=50\%)
  \item val_TBR. valleculae score given by rater TB on occasion  2 (0: no, 1: <50\%, 2: >=50\%)
  \item val_MH. valleculae score given by rater MH on occasion  1 (0: no, 1: <50\%, 2: >=50\%)
  \item val_MHR. valleculae score given by rater MH on occasion  2 (0: no, 1: <50\%, 2: >=50\%)
  \item val_CO. valleculae score given in consensus on occasion  1 (0: no, 1: <50\%, 2: >=50\%)   
  \item val_COR. valleculae score given in consensus on occasion  2 (0: no, 1: <50\%, 2: >=50\%)
}
}
