% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{fleiss.sum}
\alias{fleiss.sum}
\title{return the value of Fleiss kappa coefficient}
\usage{
fleiss.sum(data)
}
\arguments{
\item{data}{NxK dataset to compute Fleiss kappa coefficient. For each of the N subjects (row), the number of observers classifying the subject in the K-categories of the scale (columns)}
}
\value{
the value of Fleiss kappa coefficent
}
\description{
return the value of Fleiss kappa coefficient
}
\details{
This function compute Fleiss kappa coefficient
}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
