% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\docType{data}
\name{fleiss_psy}
\alias{fleiss_psy}
\title{Psychiatric diagnosis (Fleiss)}
\format{A data frame with 30 rows and 6 variables}
\description{
A dataset containing the diagnosis of depression of 50 subjects on three binary scales: diagnosis, BDI and GHQ.  The variables are as follows:
\itemize{
  \item subject. subject ID (1--30)
  \item cat1. number of observers classifying the subject in category 1 (depression)
  \item cat2. number of observers classifying the subject in category 2 (personality disorder)
  \item cat3. number of observers classifying the subject in category 3 (schizophrenia)
  \item cat4. number of observers classifying the subject in category 4 (neurosis)
  \item cat5. number of observers classifying the subject in category 5 (other psychiatric disorder)
}
}
