% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{hot.test}
\alias{hot.test}
\title{Hotelling's T square test}
\usage{
hot.test(kappa, var_kappa, a.level, NN)
}
\arguments{
\item{kappa}{a vector of length G with the coefficients to be compared}

\item{var_kappa}{the G x G variance-covariance matrix of the coefficients. It can be provided in general or obtained with the \code{delta.pair}, the \code{boot.pair}, the \code{boot.many1}, the \code{boot.many2}, the \code{delta.many1} and the \code{delta.many2} functions in the kappa context.}

\item{a.level}{the significance level}

\item{NN}{the number of clusters if the delta method is used to determine the variance-covariance matrix or the number of bootstrap iterations if the bootstrap method is used}
}
\value{
$kappa a G x 2 matrix with the coefficients in the first column and their corresponding standard error in the second column

$T_test a vector of length 2 with the value of Hotelling's T square test as first element and the corresponding p-value as second element

$confidence confidence intervals for the pairwise comparisons of the coefficients

$cor the G x G correlation matrix of the coefficients
}
\description{
This function performs Hotelling's T square test when the vector of coefficients and the variance-covariance matrix are provided
}
\details{
This function performs Hotelling's T square test to compare dependent coefficients when the vector of coefficients and the variance-covariance matrix are provided.
}
\examples{
 
vect<-c(0.3,0.4,0.5)#vector of coefficients
v_c<-matrix(c(0.01,0.005,0.007,0.005,0.015,0.006,0.007,0.006,0.05),ncol=3)#variance-covariance matrix
hot.test(kappa=vect,var_kappa=v_c,a.level=0.05,NN=50)
}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
