% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{print.kap}
\alias{print.kap}
\title{print objects of the class kap}
\usage{
\method{print}{kap}(x, ...)
}
\arguments{
\item{x}{outputs from the Hotelling's T^2 test}

\item{...}{further arguments}
}
\description{
print objects of the class kap
}
\details{
This function prints the outputs from Hotelling's T square test
}
\examples{
 
#dataset (not multilevel) (Vanbelle and Albert,2008)

set.seed(103) #to get the same results as in the paper
data(depression)
attach(depression)
a<-boot.pair(data=cbind(diag,BDI,diag,GHQ),cluster_id=ID,weight='unweighted',ITN=1000)
print(a)

}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
