% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiagree.R
\name{print.kap1}
\alias{print.kap1}
\title{print objects of the class kap1}
\usage{
\method{print}{kap1}(x, ...)
}
\arguments{
\item{x}{Kappa coefficient and 95\% confidence interval}

\item{...}{further arguments}
}
\description{
print objects of the class kap1
}
\details{
This function prints kappa coefficients and 95\% confidence intervals
}
\examples{
 
#dataset (Vanbelle and Albert, 2008)

data(depression)
attach(depression)
a<-boot.pair(data=cbind(diag,BDI),cluster_id=ID,weight="unweighted",ITN=2000)
print(a)

}
\author{
Sophie Vanbelle \email{sophie.vanbelle@maastrichtuniversity.nl}
}
